!!! | tls: OpenSSL 1.1 support follow-up
The previous commit did not fix every OpenSSL API problems
in tls.c, because it was just cherry-picked from a later
branch, and between the current tls.c and the tls.c state of
the later branch, one line was removed, which would have been
fixed with the OpenSSL 1.1 support commit.

!!! means this is a local only patch. This patch will probably
not be needed, when we upgrade FreeTDS to the next release.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>