configure/unixodbc: try to use pkg-config before odbc_config
The odbc_config has a fixed path, making it impossible to deploy the libs and use as a dependency during build.

Note: this is not a full fledged solution as it ignores a fact that user asks for no unixodbc, as in our case we always have and want it.
Signed-off-by: Kokan <kokaipeter@gmail.com>