odbc: connection-string increas option name limit from 24 to 1024
Before this, the 24 was a fine limit as the limited allowed option, but with allowing freetds.conf option to be configured, this must be increased. My current changeset only increases to yet a fixed value, it would be nicer to provide a dynamic size. But hey, 1024 is a nice large option name.

Signed-off-by: Kokan <kokaipeter@gmail.com>